+++ app/controllers/concerns/layered/assistant/stoppable_response.rb
@@ -1,0 +1,13 @@
+module Layered
+  module Assistant
+    module StoppableResponse
+      def stop
+        if @conversation.stop_response!
+          head :ok
+        else
+          head :no_content
+        end
+      end
+    end
+  end
+end