+++ config/routes.rb
@@ -7,11 +7,13 @@ Layered::Assistant::Engine.routes.draw do
     resources :models, only: [:index, :new, :create, :edit, :update, :destroy]
   end
   resources :conversations, only: [:index, :show, :new, :create, :edit, :update, :destroy] do
+    patch :stop, on: :member
     resources :messages, only: [:index, :create, :destroy]
   end
 
   namespace :panel do
     resources :conversations, only: [:index, :show, :new, :create, :destroy] do
+      patch :stop, on: :member
       resources :messages, only: [:create]
     end
   end
@@ -19,11 +21,13 @@ Layered::Assistant::Engine.routes.draw do
   namespace :public do
     resources :assistants, only: [:index, :show]
     resources :conversations, only: [:show, :create] do
+      patch :stop, on: :member
       resources :messages, only: [:create]
     end
 
     namespace :panel do
       resources :conversations, only: [:index, :show, :new, :create] do
+        patch :stop, on: :member
         resources :messages, only: [:create]
       end
     end